+++ pkg/apis/mysql/v1alpha1/mysqlcluster_types.go
@@ -95,6 +95,14 @@ type MysqlClusterSpec struct {
 	// +optional
 	BackupScheduleJobsHistoryLimit *int `json:"backupScheduleJobsHistoryLimit,omitempty"`
 
+	// BackupResources sets resources to backup pod container
+	// +optional
+	BackupResources core.ResourceRequirements `json:"backupResources,omitempty"`
+
+	// BackupLabels set labels to backup pods
+	// +optional
+	BackupLabels map[string]string `json:"backupLabels,omitempty"`
+
 	// A map[string]string that will be passed to my.cnf file.
 	// +optional
 	MysqlConf MysqlConf `json:"mysqlConf,omitempty"`
@@ -371,7 +379,6 @@ type MysqlClusterStatus struct {
 // +kubebuilder:printcolumn:name="Replicas",type="integer",JSONPath=".spec.replicas",description="The number of desired nodes"
 // +kubebuilder:printcolumn:name="Age",type="date",JSONPath=".metadata.creationTimestamp"
 // +kubebuilder:resource:shortName=mysql
-//
 type MysqlCluster struct {
 	metav1.TypeMeta   `json:",inline"`
 	metav1.ObjectMeta `json:"metadata,omitempty"`
@@ -382,7 +389,6 @@ type MysqlCluster struct {
 
 // MysqlClusterList contains a list of MysqlCluster
 // +kubebuilder:object:root=true
-//
 type MysqlClusterList struct {
 	metav1.TypeMeta `json:",inline"`
 	metav1.ListMeta `json:"metadata,omitempty"`