+++ CHANGELOG.md
@@ -10,6 +10,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 * Add `KeepAfterDelete` in `.Spec.VolumeSpec` to keep pvc after mysql cluster been deleted.
 * Add default resource to init container.
+* Add `MysqlCluster.Spec.BackupResources` to allow the user specifying resources for backup job container.
+* Add `MysqlCluster.Spec.BackupLabels` to allow the user specifying labels for backup job pods.
 
 ### Changed
 ### Removed